ops(heroku):remove cache clean on local

+++ scripts/reduce-slug-size-heroku.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+[ -z "${LUMIUM_HEROKU_BUILD}" ] && exit 0;
 echo "cleaning up..."
 rm -rf lumium-renderer/target
 rm -rf clang+llvm-14.0.0-x86_64-linux-gnu-ubuntu-18.04
